refactor: コードベースの大半をsharedに切り出し
