+++ packages/extension/entrypoints/options/components/constants.ts
@@ -5,4 +5,4 @@ export {
   STORAGE_KEY,
   HISTORY_STORAGE_KEY,
   SUBJECTS_STORAGE_KEY,
-} from "@mikoto-moocs/shared";
+} from "@mikoto-moocs-sharp/shared";
