+++ packages/shared/src/components/schedule/index.ts
@@ -1,0 +1,4 @@
+export * from "./ScheduleEditor";
+export * from "./ScheduleGrid";
+export * from "./ScheduleCell";
+export * from "./ClassEditModal";
