+++ packages/shared/src/hooks/index.ts
@@ -1,2 +1,6 @@
 export * from "./useCountdown";
 export * from "./useCurrentTime";
+export * from "./dom/useElementObserver";
+export * from "./dom/useTextareaObserver";
+export * from "./schedule/useSchedule";
+export * from "./schedule/useScheduleHistory";
