+++ packages/shared/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./dom";
 export * from "./text";
+export * from "./subjectExtractor";
